requirements.txt
agno 
python-dotenv
fastapi[all]
groq
duckduckgo-search
google-genai
duckdb
openai
twilio
firecrawl